refactor: 숲꾸 탭의 카드 컴포넌트 재구성
