+++ src/app/forest/page.tsx
@@ -1,7 +1,21 @@
-import FeatureGrid from "./(components)/feature-grid";
 import Ranking from "./(components)/ranking";
 import GoForest from "./(components)/go-forest";
 import EventSliderBanner from "./(components)/event-slider-banner";
+import { PATH } from "@/lib/paths";
+import FeatureCard from "./(components)/feature-card";
+
+const menuTabInfo = [
+  {
+    title: "상점",
+    href: PATH.HOME,
+    thumbnail: "/illusts/3d-forest.svg",
+  },
+  {
+    title: "출석체크",
+    href: PATH.HOME,
+    thumbnail: "/illusts/3d-forest.svg",
+  },
+];
 
 export default function ForestPage() {
   return (
@@ -10,7 +24,11 @@ export default function ForestPage() {
       <EventSliderBanner />
 
       {/** 상점, 출석체크 */}
-      <FeatureGrid />
+      <div className="flex gap-4">
+        {menuTabInfo.map((item, idx) => (
+          <FeatureCard key={`${item.title}_${idx}`} {...item} />
+        ))}
+      </div>
 
       {/** 명예의 전당 */}
       <Ranking />
